Host compute service failure alarm removal
Removal of stale alarm 270.001(Host compute service failure)
is raised by the vim. This might be an old reference to nova.
It’s likely not in use since stx.

Test Plan:
PASS: Verify with a load without the changes (removal of alarm)
and the event log in platform.log shows an entry for 270.001 alarm.
PASS: Verify with a load with changes of alarm removal and
the event log in platform.log does not show an entry for 270.001 alarm.

Depends-On: https://review.opendev.org/c/starlingx/fault/+/872603

Closes-Bug: 2004744

Change-Id: Icafb079fc2b58fb4126ac325804901ebd3f8f66e
Signed-off-by: Vanathi.Selvaraju <vanathi.selvaraju@windriver.com>