Change pod rollout timeout for K8s root CA change
The Puppet timeout for pod rollout in stages "trustbothcas" and
"trustnewca" was recently changed from 600s to 3600s. In this commit,
the timeouts for these stages in the kubernetes root CA update strategy
are also updated.

Test Plan:

PASS: In a AIO-SX, execute the Kubernetes Root CA update through
sw-manager strategy and check that it is completed successfully without
reaching any timeout in stages "trustbothcas" and "trustnewca".

PASS: In a AIO-SX, artificially make the pod rollout script hang for 15
minutes and check that the stages "trustbothcas" and "trustnewca" are
still completed successfully.

Closes-Bug: 2004594
Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: I96d04de95e424e15bd79f049be644909bb0dcff7