NFVI: Default guest_plugin_disabled to True
Following the work previously done in [1] and [2] to deactivate
guest-related services in VIM, this commit changes the default value of
config variable `guest_plugin_disabled` to `True` so that it reflects
the change proposed in [3] (same modification but on Puppet's side).

As reported in [3], loading this plugin while having some of its
services deactivated (or functionalities removed) has proven to be
a problem when stx-openstack is applied, as both nova-compute service
and hypervisor are caught in an enable/disable loop indefinitely after
the first host lock/unlock with the application applied.

[1] https://review.opendev.org/c/starlingx/nfv/+/869817
[2] https://review.opendev.org/c/starlingx/nfv/+/870538
[3] https://review.opendev.org/c/starlingx/stx-puppet/+/879359

Test Plan (on AIO-SX):
PASS - Remove 'guest_plugin_disabled' from /etc/nfv/vim/config.ini,
       reload VIM services and verify that the guest plugin wasn't
       loaded by default:
       $ tail -f /var/log/nfv*.log

Related-Bug: 2015088

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/879359

Signed-off-by: Luan Nunes Utimura <LuanNunes.Utimura@windriver.com>
Change-Id: I7e254fe2db2a6bcc6b98a26cc712c5d03ef7ffad