Add logs to inform when nodes are tainted
Add logs to inform when the 'services=disabled:NoExecute' taint is added
to nodes as well as when they are removed.

This aims to improve future log analysis by facilitating the
identification of cases where taints could be mistakenly added to nodes.

Test Plan:
PASS: lock controller and check if the operation is properly logged
PASS: unlock controller and check if the operation is properly logged

Partial-Bug: 2022008
Change-Id: Ie9c7432211621a9fbf7aa90282dfb91405f90c33
Signed-off-by: Igor Soares <igor.piressoares@windriver.com>