Optimize kubelet upgrade phase using VIM orchestrator.
Optimizing kubelet upgrade by reducing the wait
time and adding in more logs for debugging.The
wait time is removed as we have the additional
kubelet state check in place (Closes-Bug: 2044209)
and an existing timeout of 900secs.

Test Plan:
PASSED: On a DX system, removed the wait time of 60 sec
and tested 3 consecutive k8s upgrade thrice.
PASSED: On a DX system, tweaked the code and tested
on the kubelet upgrade retry timeout(900sec).
PASSED: On a DX system, tweaked code and tested
the existing behaviour that retry is not
occuring in case of failure.

Closes-Bug: 2045776

Change-Id: I321d7eae5ef7ebd29c1c6aca97992e3e20acb457
Signed-off-by: Vanathi.Selvaraju <vanathi.selvaraju@windriver.com>