Alarm 900.701 raised on failing to remove node taint.
Additional checks added to ensure that alarm is
raised by VIM on failing to remove node taint.

Test Plan:
PASSED: On a DX system, locked and unlocked one of
the controller to check if taints are removed.
PASSED: On a DX system, tweaked the code to
fail untainting of node.
Alarm 900.701 is raised as there is node taint.
PASSED: On a DX system, check if the alarm 900.701
is removed on locking the node.
PASSED: Deployed a DX system with ISO that has the
changes. No trace of alarm 900.701 after
bootstrapping.
PASSED: On a DX system Node taint alarm exists with
multiple taints, node was locked followed by unlock.
On successful untaint, the alarm is cleared.
PASSED: On a DX system Node taint alarm exists, node was
locked followed by unlock. On successful untaint, the
alarm 900.701 is cleared.

Closes-Bug: 2046273
Depends-On: https://review.opendev.org/c/starlingx/fault/+/904788

Change-Id: I4206b336cbe0021f2b45e3b3cd24b42ca43bc60e
Signed-off-by: Vanathi.Selvaraju <vanathi.selvaraju@windriver.com>