Fix: Do not consider locked hosts as deployed
When reentering a sw-deploy-strategy it is possible to have a host in
deployed state while still being locked.  This is not considered an
acceptable state by USM and must be addressed before activation.

This change enables VIM to detect and remediate this.  This was done
by improving the logic that decided when and which steps are skipped.

TEST PLAN
PASS: AIO-SX sw-deploy-strategy (major)
PASS: AIO-SX sw-deploy-strategy (major)
* Manually do sw-deploy get to host-deployed while locked
* Run VIM after, VIM should complete sw-deploy
PASS: AIO-DX sw-deploy-strategy (major)
PASS: AIO-DX sw-deploy-strategy (major)
* Manually do sw-deploy get to host-deployed (c1) while locked
* Run VIM after, VIM should complete sw-deploy
PASS: AIO-DX sw-deploy-strategy (major)
* Manually do sw-deploy get to host-deployed (c0) while locked
* Run VIM after, VIM should complete sw-deploy

Story: 2011045
Task: 50920
Change-Id: I8e0ade8892881ccef71fb35fb4a361803f22792d
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>