Call get_tenants only with openstack installed
This change aims to reduce the number of calls to /projects
against the platform's keystone service. With this change get_tenant
will only be called when openstack is installed, since in this case
it's expected that users will create different projects to isolate their
resources.

This reduces the number of requests to keystone from ~ 286 to 2 per hour

Test case:
PASS: AIO-SX and AIO-DX: sw-manager kube-rootca-update-strategy
PASS: DC: dcmanager kube-rootca-update-strategy
PASS: Verify /var/log/nfv-vim*.log and sysinv*.log for errors
PASS: Apply openstack and verify that requests are sent to the
      containerized keystone.
PASS: openstack with deployed VMs to trigger migration or stop/start
      steps with the nova caveat as mentioned in comments by
      Thales.

Closes-Bug: 2081365
Change-Id: I9d002a755330ff4ccf46bdc90dc332ba38ce5512