sw-deploy-strategy apply error message improvements.
sw-deploy-strategy apply throws error when apply is attempted
before create strategy is completed. The current error is not
clear. As part of this fix, the error messages are improved.

Test Plan:
PASSED: When create strategy is in progress, apply strategy is attempted.
PASSED: When apply strategy is in progress, apply is attempted
again
PASSED: When create strategy failed, apply strategy is attemped.

Closes-Bug: 2082871

Change-Id: If78b3040edcd7d8c60a0d53d8c4535164ce5cc5d
Signed-off-by: Vanathi.Selvaraju <vanathi.selvaraju@windriver.com>