Fix: Rentrancy during sw-deploy downgrades
Reentrancy is not working for downgrades because it is checking
information on the from-release instead of the to-release.

The fix is to check for a deployment in progress rather than check for
specific states.

TEST PLAN
PASS: AIO-DX sw-deploy-strategy upgrade
PASS: AIO-DX sw-deploy-strategy upgrade
* Reenter on start-done
PASS: AIO-DX sw-deploy-strategy downgrade
PASS: AIO-DX sw-deploy-strategy downgrade
* Reenter on start-done

Closes-Bug: 2089804
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>
Change-Id: I751e0063ac9dff5e60d22a64fa754262e944159e