Alarm 750.006 alarm takes too long to clear up
Alarm 750.006 is waiting for vim operation to
complete and vim waits for 750.006 to get cleared
creating a deadlock case. The current patch strategy
has a timeout of 30mins to clear the alarm if exists.
The timeout has been reduced to 2mins to avoid
longer wait time.

Test Plan:
PASSED: Verified patch update, after 2mins
the alarm is ignored and vim operation continues.
PASSED: Verified that patch update continues,
when configuration is reapplied and alarm 750.006
is cleared.

Closes-Bug: 2091315

Change-Id: I050e424c65fa741d8807fc813aced6f060f80286
Signed-off-by: Vanathi.Selvaraju <vanathi.selvaraju@windriver.com>