sw-deploy strategy UpgradeActivate retry timeout changes
The retry timeout of Upgrade Activate is changed as
part of this fix from 120 secs to 30 secs(this value
is considered after testing).

Before this change:
VIM orchestration currently waits for 6mins on
activation failure (i.e. 3mins for each retry,
there are 2 retries) which adds to the overall upgrade time.

After this change:
The timeout would be reduced by 1min 15~30 secs
(around 4mins 30secs).The number of retry is increased to 3
and with each retry the delay is increased by 30secs.

TEST PLAN
PASSED: Activation script failed on DX system,
        VIM orchestration retry done with
        a delay of 38secs(alarm audit takes 30-35 sec)
PASSED: With the configuration changes on DX system,
        no failure with activation script, upgrade
        is successful.
PASSED: Upgrade from 24.09 to 25.09, with activation
        failure induced to check retries.

Story: 2011045
Task: 51650

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/940046

Change-Id: Id343a8edd50e19ea5fa37378b9452480476a6d55
Signed-off-by: Vanathi.Selvaraju <vanathi.selvaraju@windriver.com>