Increase deploy delete timeout
Currently deploy delete phase uses default timeout of 60s,
which is not enough now that the upgrade scripts are executed
with action 'delete' during this phase.

This commit increases the timeout of this step to 5 minutes.

Test Plan
PASS: create sw-deploy-strategy and verify the timeout set
      for deploy delete phase is set to 5 minutes/300 seconds
PASS: e2e stx-10 to stx-11 orchestrated upgrade

Partial-bug: 2119671

Change-Id: I68450569595b99494759c5f93552c51c06da2ff9
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>