[nfvi-plugins] Improves IP Familly detection logs
[1] added the IPv6 support for the REST API base implementation shared
by some of the NFVI plugins (Infrastructure and Compute REST APIs).
However, during the debug of the launchpad (LP) #2121390, it was noticed
that some logs added by [1] cause confusion for APIs listening on IPv4
addresses. For example, the following message is currently displayed for
the Infrastructure REST API listening on local IP "127.0.0.1":

    "Cannot transform IP 127.0.0.1 to IPv6. illegal IP address string
     passed to inet_pton".

This message gives the wrong idea that IP would always be expected as an
IPv6 address.

This change improves this log by displaying the IP and address family
configured for each of the APIs, e.g.:

    "REST API setup to listen on IPv6 address xxxx:xx:xx:xxx::x"
    "REST API setup to listen on IPv4 address 127.0.0.1"

Furthermore, the change asserts that the IP configured should be either
detected as IPv4 or IPv6 by "socket.inet_pton" calls (the "host" config
supports valid IP addresses, it doesn't support names like "localhost").

Test Plan (performed on IPv6 AIO-DX and IPv4 standard deployments):
[PASS] Upload and apply stx-openstack app
[PASS] Infrastructure REST API loaded and IP detection logs displayed in
       the expected format (nfv-vim.log)
[PASS] Compute REST API loaded and IP detection logs displayed in the
       expected format (nfv-vim.log)

[1] https://review.opendev.org/c/starlingx/nfv/+/955947

Related-Bug: #2121390

Change-Id: Ic9c468a71c14e1bde1cb56c1d33373ba1089c0fe
Signed-off-by: Alex Figueiredo <alex.fernandesfigueiredo@windriver.com>